Ph5 robotics: fix the flight-test defects and hermetize the search tests
415/415 game tests on 26.1.2 and 26.2; unit tests green on 26.1.2,
26.2 and 1.21.1. The two step-5 failures were three real defects
plus one test-world bug:

- AIRobotSearchAndGotoBlock's 3-arg ctor defaulted maxDistanceToEnd
  to 64 where 7.1.x's board-facing default is 0. With a positive
  radius, PathFinding.endReached fires at the first soft cell in
  range, so the search reports a two-cell stub path and the goto leg
  flies the robot nowhere. 0 ends the pathfind on the target cell,
  yielding the real approach path (the target cell itself is dropped
  by the search, as 7.1.x did).
- AIRobotGoto's movement targets (and AIRobotGotoBlock's final
  position) were float. The 26.x world is ±30M blocks, where float's
  ulp is 1.0, so the +0.5 cell-centre offset is silently dropped at
  game-test arena coordinates (|z| ~ 1.4e7) and the robot aims half
  a block off. In-memory fields are now double; NBT stays float.
- The search tests were not hermetic: the search scans 64 blocks,
  and in the shared minecraft:default game-test world a neighbouring
  test's leftover oak log (two cells away) is a legal, closer target
  than the log the test places. Both now run in private
  test_environments (the codebase's established pattern).
- The world-properties wheat pin was a test-world bug, not a
  property bug: vegetation blocks re-validate canSurvive whenever a
  neighbour changes (VegetationBlock.updateShape), and the dark
  underground arena floor (bedrock) fails the check, so the crops
  turned to air before the read and the assertFalse pins passed for
  the wrong reason. The test now lays farmland/soul sand/dirt under
  the plants and a glowstone for wheat's light-8 requirement.

The AIRobot no-terminated-flag re-cycle (cycle() re-runs update()
after terminate()) is faithful to 7.1.x and required by production
(EntityRobot.tick cycles mainAI forever; boards never
self-terminate) — deliberately not changed.

+++ src/main/java/buildcraft/robotics/ai/AIRobotGoto.java
@@ -17,14 +17,17 @@ import buildcraft.api.robots.IRobotAccess;
  *  when the move is done. */
 public abstract class AIRobotGoto extends AIRobot {
 
-    protected float nextX, nextY, nextZ;
+    // Double, not float: the 26.x world is ±30M blocks, and float's ulp is already 1.0 past |8.4M| — a
+    // cell-centre offset of +0.5 is silently dropped at that scale and the robot aims half a block off
+    // (the game-test arenas sit at |z| ≈ 1.4e7, where this bit the search-and-goto flight).
+    protected double nextX, nextY, nextZ;
     protected double dirX, dirY, dirZ;
 
     public AIRobotGoto(IRobotAccess iRobot) {
         super(iRobot);
     }
 
-    protected void setDestination(IRobotAccess robot, float x, float y, float z) {
+    protected void setDestination(IRobotAccess robot, double x, double y, double z) {
         nextX = x;
         nextY = y;
         nextZ = z;
@@ -45,7 +48,7 @@ public abstract class AIRobotGoto extends AIRobot {
             dirZ = 0;
         }
 
-        robot.setDeltaMovement(new net.minecraft.world.phys.Vec3(dirX / 10F, dirY / 10F, dirZ / 10F));
+        robot.setDeltaMovement(new net.minecraft.world.phys.Vec3(dirX / 10D, dirY / 10D, dirZ / 10D));
     }
 
     @Override
